Use 'type' instead of 'which' to check for ack
Use the 'type' built-in command to check for ack or ack-grep instead of
the external 'which' command. This should offer some speed improvements
since we're not calling an external program. Also check for the
existence of ack or ack-grep before defining the function and return
with an error if one of them isn't found.